common/cuda: Move opal_datatype_cuda functionality into common cuda
