+++ ompi/mca/mtl/base/mtl_base_datatype.h
@@ -28,7 +28,6 @@
 #if OPAL_CUDA_SUPPORT
 #include "opal/mca/common/cuda/common_cuda.h"
 #include "opal/datatype/opal_convertor.h"
-#include "opal/datatype/opal_datatype_cuda.h"
 #endif
 
 #ifndef MTL_BASE_DATATYPE_H_INCLUDED
