+++ ompi/mca/mtl/ofi/mtl_ofi.h
@@ -50,7 +50,6 @@
 #if OPAL_CUDA_SUPPORT
 #include "opal/mca/common/cuda/common_cuda.h"
 #include "opal/datatype/opal_convertor.h"
-#include "opal/datatype/opal_datatype_cuda.h"
 #endif
 
 BEGIN_C_DECLS
