+++ opal/datatype/opal_datatype_copy.c
@@ -72,7 +72,7 @@ static size_t opal_datatype_memop_block_size = 128 * 1024;
 #include "opal_datatype_copy.h"
 
 #if OPAL_CUDA_SUPPORT
-#include "opal_datatype_cuda.h"
+#include "opal/mca/common/cuda/common_cuda.h"
 
 #undef MEM_OP_NAME
 #define MEM_OP_NAME non_overlap_cuda
